pyproject.toml
[tool.poetry]
name = "snakeclef"
version = "0.1.0"
description = ""
authors = ["Jack Etheredge <jack.etheredge@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.2"
torch = "^2.3.0"
torchvision = "^0.18.0"
torchaudio = "^2.3.0"
scikit-learn = "^1.4.2"
imbalanced-learn = "^0.12.2"
hydra-core = "^1.3.2"
tqdm = "^4.66.4"
focal-loss-torch = "^0.1.2"
torch-lr-finder = "^0.2.1"
timm = "^0.9.16"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
